# Backend/Swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpSchemaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/signin": {
      "post": {
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInSchemaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout/{id}": {
      "get": {
        "operationId": "AuthController_logout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/get-otp/{email}": {
      "get": {
        "operationId": "AuthController_getOTP",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "path",
            "description": "User Email",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/refresh/{id}": {
      "patch": {
        "operationId": "AuthController_refreshTokens",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshSchemaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateSchemaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
...
    "/category": {
      "get": {
        "operationId": "CategoryController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Category"
        ]
      },
      "post": {
        "operationId": "CategoryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/category/{id}": {
      "get": {
        "operationId": "CategoryController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Category"
        ]
      },
      "patch": {
        "operationId": "CategoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Category"
        ]
      },
      "delete": {
        "operationId": "CategoryController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
...
  },
  "info": {
    "title": "ExBuySell - Api Documentation",
    "description": "",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "SignUpSchemaDTO": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "minLength": 4
          },
          "last_name": {
            "type": "string",
            "minLength": 3
          },
          "password": {
            "type": "string",
            "minLength": 4
          },
          "confirm_password": {
            "type": "string",
            "minLength": 4
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 4
          }
        },
        "required": [
          "first_name",
          "last_name",
          "password",
          "confirm_password",
          "email"
        ]
      },
      "SignInSchemaDTO": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 4
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 4
          }
        },
        "required": [
          "password",
          "email"
        ]
      },
      "RefreshSchemaDTO": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 4
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "UserUpdateSchemaDTO": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "minLength": 4
          },
          "last_name": {
            "type": "string",
            "minLength": 4
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "USER",
              "MONITER"
            ]
          },
          "businessName": {
            "type": "string",
            "nullable": true
          },
          "contactName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "permissions": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "zip_code": {
            "type": "integer",
            "format": "int64"
          },
          "background": {
            "type": "string"
          },
          "is_online": {
            "default": false,
            "type": "boolean"
          },
          "password_hash": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string",
            "nullable": true
          },
          "otp_code": {
            "type": "string"
          },
          "approved": {
            "default": false,
            "type": "boolean"
          },
          "is_email_verified": {
            "default": false,
            "type": "boolean"
          },
          "is_phone_verified": {
            "default": false,
            "type": "boolean"
          }
        }
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "image_path": {
            "type": "string",
            "minLength": 4,
            "nullable": true
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}